+++ framework/Database/Migrations/2023_12_29_085321_create_migrations_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
             'CREATE TABLE migrations (' .
             'id INT auto_increment,' .
             'name VARCHAR(255) NOT NULL,' .
+            'createdAt TIMESTAMP DEFAULT CURRENT_TIMESTAMP,' .
+            'updatedAt DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,' .
             'PRIMARY KEY (id)' .
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;'
         );