+++ framework/Database/Migrations/2023_12_29_100125_create_users_table.php
@@ -12,9 +12,11 @@ return new class extends Migration
             'id INT auto_increment,' .
             'firstname VARCHAR(30) NOT NULL,' .
             'lastname VARCHAR(30) NOT NULL,' .
-            'login VARCHAR(30) NOT NULL,' .
+            'username VARCHAR(30) NOT NULL,' .
             'password VARCHAR(255) NULL,' .
-            'forcePwdChange tinyint(1) NOT NULL,' .
+            'isPwdChangeForced tinyint(1) NOT NULL,' .
+            'createdAt TIMESTAMP DEFAULT CURRENT_TIMESTAMP,' .
+            'updatedAt DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,' .
             'PRIMARY KEY (id)' .
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;'
         );